refactor: Seat 관련 클래스 패키지 위치 변경 및 통합
- service.schedule의 SeatRepository 기능을 service.ticketing으로 통합
- Seat 엔티티 참조 패키지 경로를 service.ticketing으로 수정
- 불필요한 ShowSchedule 및 ShowSectionPrice 엔티티 삭제
